# app/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Transaction Engine API",
        "description": "API para processamento assíncrono de transações financeiras",
        "version": "1.0.0"
    },
    "paths": {
        "/api/accounts/{id}/balance": {
            "get": {
                "tags": [
                    "Accounts"
                ],
                "summary": "Consulta saldo da conta",
                "operationId": "c1810a06cff78bccfc24a3554a105468",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Saldo da conta",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "balance": {
                                            "type": "number",
                                            "example": 1000
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Conta não encontrada"
                    }
                }
            }
        },
        "/api/transactions": {
            "post": {
                "tags": [
                    "Transactions"
                ],
                "summary": "Cria uma nova transação",
                "operationId": "ab765dced58cfe2f620929db31452145",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "source_account_id",
                                    "amount",
                                    "type"
                                ],
                                "properties": {
                                    "source_account_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "destination_account_id": {
                                        "type": "integer",
                                        "example": 2
                                    },
                                    "amount": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 100.5
                                    },
                                    "type": {
                                        "type": "string",
                                        "enum": [
                                            "credit",
                                            "debit",
                                            "transfer"
                                        ]
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "202": {
                        "description": "Transação criada com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "transaction_id": {
                                            "type": "integer"
                                        },
                                        "status": {
                                            "type": "string",
                                            "example": "PENDING"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Erro de validação"
                    }
                }
            }
        }
    },
    "tags": [
        {
            "name": "Accounts",
            "description": "Accounts"
        },
        {
            "name": "Transactions",
            "description": "Transactions"
        }
    ]
}